+++ cvat/apps/engine/admin.py
@@ -4,16 +4,26 @@
 # SPDX-License-Identifier: MIT
 
 from django.contrib import admin
-from .models import Task, Segment, Label, AttributeSpec
+from .models import Task, Segment, Job, Label, AttributeSpec
+
+class JobInline(admin.TabularInline):
+    model = Job
+    can_delete = False
+
+    # Don't show extra lines to add an object
+    def has_add_permission(self, request, object=None):
+        return False
 
 class SegmentInline(admin.TabularInline):
     model = Segment
+    show_change_link = True
     readonly_fields = ('start_frame', 'stop_frame')
     can_delete = False
 
-    # Don't show on admin index page
+    # Don't show extra lines to add an object
     def has_add_permission(self, request, object=None):
         return False
+
 
 class AttributeSpecInline(admin.TabularInline):
     model = AttributeSpec
@@ -35,14 +45,23 @@ class LabelAdmin(admin.ModelAdmin):
         AttributeSpecInline
     ]
 
+class SegmentAdmin(admin.ModelAdmin):
+    # Don't show on admin index page
+    def has_module_permission(self, request):
+        return False
+
+    inlines = [
+        JobInline
+    ]
 
 class TaskAdmin(admin.ModelAdmin):
     date_hierarchy = 'updated_date'
     readonly_fields = ('size', 'path', 'created_date', 'updated_date',
         'overlap', 'flipped')
-    list_display = ('name', 'mode', 'owner', 'created_date', 'updated_date')
+    list_display = ('name', 'mode', 'owner', 'assignee', 'created_date', 'updated_date')
     search_fields = ('name', 'mode', 'owner__username', 'owner__first_name',
-        'owner__last_name', 'owner__email')
+        'owner__last_name', 'owner__email', 'assignee__username', 'assignee__first_name',
+        'assignee__last_name')
     inlines = [
         SegmentInline,
         LabelInline
@@ -54,4 +73,5 @@ class TaskAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Task, TaskAdmin)
+admin.site.register(Segment, SegmentAdmin)
 admin.site.register(Label, LabelAdmin)