+++ cvat/settings/base.py
@@ -54,7 +54,8 @@ INSTALLED_APPS = [
     'cacheops',
     'sendfile',
     'dj_pagination',
-    'revproxy'
+    'revproxy',
+    'rules'
 ]
 
 if 'yes' == os.environ.get('TF_ANNOTATION', 'no'):
@@ -99,6 +100,18 @@ TEMPLATES = [
 ]
 
 WSGI_APPLICATION = 'cvat.wsgi.application'
+
+# Django Auth
+DJANGO_AUTH_TYPE = 'BASIC'
+LOGIN_URL = 'login'
+LOGIN_REDIRECT_URL = '/'
+AUTH_LOGIN_NOTE = '<p>Have not registered yet? <a href="/auth/register">Register here</a>.</p>'
+
+AUTHENTICATION_BACKENDS = [
+    'rules.permissions.ObjectPermissionBackend',
+    'django.contrib.auth.backends.ModelBackend'
+]
+
 
 # Django-RQ
 # https://github.com/rq/django-rq